+++ core/src/commonMain/kotlin/extensions/Node.kt
@@ -6,18 +6,18 @@ import dev.inmo.navigation.core.NavigationChain
 import dev.inmo.navigation.core.NavigationNode
 import kotlinx.coroutines.flow.*
 
-val <T> NavigationNode<T>.onChainsStackDiffFlow
+val <Base> NavigationNode<out Base, Base>.onChainsStackDiffFlow
     get() = flow {
-        val previous = mutableListOf<NavigationChain<T>>()
+        val previous = mutableListOf<NavigationChain<Base>>()
         emit(previous.applyDiff(subchainsFlow.value, strictComparison = true))
         subchainsFlow.collect {
             val newValue = subchainsFlow.value
             emit(previous.applyDiff(newValue, strictComparison = true))
         }
     }
-val <T> NavigationNode<T>.onChainAddedFlow
+val <Base> NavigationNode<out Base, Base>.onChainAddedFlow
     get() = onChainsStackDiffFlow.map { it.added }.filter { it.isNotEmpty() }
-val <T> NavigationNode<T>.onChainRemovedFlow
+val <Base> NavigationNode<out Base, Base>.onChainRemovedFlow
     get() = onChainsStackDiffFlow.map { it.removed }.filter { it.isNotEmpty() }
-val <T> NavigationNode<T>.onChainReplacedFlow
+val <Base> NavigationNode<out Base, Base>.onChainReplacedFlow
     get() = onChainsStackDiffFlow.map { it.replaced }.filter { it.isNotEmpty() }